attempt to refactor module
